Implement time series storage for door state.
The door status is being committed to a Cassandra database in reverse
temporal order (starting from ~0UL) and the hash of data being
transmitted is verified before it is learned into the database.

This is somewhat vulnerable to replay attacks of course, but then again
we don’t care too much about them since they would just overwrite the
same value. They could be used for a write load DoS though, so perhaps
this should be addressed later as well. An idea for that would be to
add an in-memory requirement that new timestamps must be higher than
old ones.